Refactor to extract data fetching from component
